% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_var.R
\name{sign_vars}
\alias{sign_vars}
\title{This function returns the predictor variables that are statistically significant in a linear model.}
\usage{
sign_vars(dependents = data.frame(), independent = vector(), alpha = 0.05)
}
\arguments{
\item{alpha}{A numeric value between 0 and 1 to be the p.value threshold}

\item{x}{A vector of the independent variables}
}
\value{
A vector identifying which variables that have a p value < alpha in in a linear model
}
\description{
This function returns the predictor variables that are statistically significant in a linear model.
}
\examples{
sign_vars(iris[,1:3], iris[,4])) #error = TRUE
sign_vars(iris[,1:3], as.vector(iris[,4]))
}
